Add Django 1.2 compatibility and fix uturn_param tag
Added RequestFactory from django-cms to allow the tests to run with
Django 1.2. Meanwhile fixed a bug in the uturn_param tag and made it
compatible with Django 1.2 by requiring the request to be passed in as
a parameter rather than relying on the takes_context argument which
didn't exist back then. Tests for the uturn_param tag have been added
as well.